Curso atribute Duracao correted from Int to Double

+++ Entities/Curso.cs
@@ -6,7 +6,7 @@
         public string Nome { get; set; }
         public string Descricao { get; set; }
         public int Nivel { get; set; }
-        public double Duracao { get; set; }
+        public int Duracao { get; set; }
         public int ProfessorId { get; set; }
         public DateTime DataInicio { get; set; }
         public double Valor { get; set; }
